+++ src/com/google/gwt/widgetideas/graphics/client/impl/GradientFactoryImplDefault.java
@@ -25,13 +25,13 @@ import com.google.gwt.widgetideas.graphics.client.GradientFactory;
  */
 public class GradientFactoryImplDefault implements GradientFactory {
 
-  public CanvasGradient createLinearGradient(float x0, float y0, float x1,
-      float y1, Element c) {
+  public CanvasGradient createLinearGradient(double x0, double y0, double x1,
+      double y1, Element c) {
     return new LinearGradientImplDefault(x0,y0,x1,y1,c);
   }
 
-  public CanvasGradient createRadialGradient(float x0, float y0, float r0,
-      float x1, float y1, float r1, Element c) {
+  public CanvasGradient createRadialGradient(double x0, double y0, double r0,
+      double x1, double y1, double r1, Element c) {
     return new RadialGradientImplDefault(x0,y0,r0,x1,y1,r1,c); 
   }
 }
